11/10/2021

+++ pages/api/categories/[id].js
@@ -1,5 +1,6 @@
 import connectDB from '../../../utils/connectDB';
 import Categories from '../../../models/categoriesModel';
+import Products from '../../../models/productModels';
 import auth from '../../../middleware/auth';
 
 connectDB();
@@ -44,6 +45,11 @@ const deleteCategory = async (req, res) => {
       return res.status(400).json({ err: 'Authentication is not valid.' });
 
     const { id } = req.query;
+
+    const products = await Products.findOne({category:id})
+
+    if(products) return res.status(400).json({ err: 'Please delete all products with a relationship' });
+
 
     await Categories.findByIdAndDelete(id);
     res.json({ msg: 'Success! Delete a  category' });
